Split tests for transforms v2 and prototype (#7278)
